fix documentation + behavior of allowNone
The docs said that when `allowNone` was true, validation would be
skipped if the value was null or undefined. What was actually happening
is that validation would not be skipped at all in those cases.

Instead, only if `allowNone` was false, the validation would fail if the
value was null/undefined.

This commit changes the behavior of this validator to actually match the
documentation. As a result, existing users may no longer see validation
errors where they previously did.